[BREAKING] ae.utils.graphics: Introduce an image coordinate/size type
Introduce a properly defined type which holds some position/size
across an axis in an image, or a derivative thereof.

Rationale:

- Using a type with a different size than size_t for a very similar
  concept caused friction due to casts, or code compiling on some
  target bitness but not another.

Changes required:

- Update uses of `int` for coordinates / image sizes to use the new
  `xy_t` type instead.

It's possible that the type change may be reverted back to `int` (e.g.
if a performance regression is observed). The `xy_t` alias will,
however, remain, and should be used for semantics.