ae.sys.net.package: Add Network.httpRequest
Stop beating around the bush with poorly-defined "high-level"
primitives such as "urlOK" / "post" / "resolveRedirect", and make the
lowest-level primitive for HTTP requests part of the public interface.
This allows Network consumers to build their own primitives on top.
For example, this will allow removing some low-level interface
extensions in CachedCurlNetwork.

The remaining methods are still useful for non-HTTP requests (when the
implementation supports it).