ae.sys.d.manager: Run GNU make as "make" on non-FreeBSD
The previous change, 8bad002a2e70f4aa1af6413a5cb4b28bc05b6770, was
incorrect in that it assumed that on POSIX, GNU make will always be
available as a command named "gmake". This assumption was incorrect;
it is only installed as "make" (at least on Arch Linux and Ubuntu).