ae.sys.git: Revert "Deprecate calling getHistory with extraRefs"
This reverts commit 4d89329925a18e558ea90a0ba270c430505abe38.

This is useful when we want to also query the history of dangling
commits (e.g. before a ref is updated to point at them, or after a ref
is updated to point away from them). Though, the terminology is
misleading in this case, as we are not providing refs (which are
already included) but commit hashes.