.github/workflows/test.yml: Use DMD 2.097 for macos-10.15
The TLS fix was merged in 2.087.1, but for some reason 2.087.1 still
fails with the same problem. 2.088.0 works, but everything crashes
when an exception (to be caught later) is thrown.

The current latest version (2.097.0) does seem to work, so just use
that, as there isn't really a point in trying to find a minimal
working version considering all the headache surrounding this
questionable platform.