ae.sys.timing: Deprecate Duration-based methods
A duration is not the best way to represent a timer task, because it
is relative to a point in time which is not specified.  Consider the
situation when the desired condition is not a duration (starting with
some interpretation of "now") elapsing, but a specific point in time;
with a duration-based API, it would need to be converted to a duration
first, and only then converted into a point in time internally in the
Timer's implementation.

The duration is also part of the TimerTask state, and is used for
restarting tasks. However, because callers generally are aware of what
duration their tasks are, and this duration is usually some hard-coded
value, this state is generally redundant.

This commit deprecates the duration state for the above reason, and
also the duration APIs, to force callers to be explicit about the
origin point in time from which any durations should be based
on.  (For example, a periodic timer may want to schedule itself to run
either the same duration after its last activation, plus any errors;
or on a strict schedule.)