ae.net.asockets: Centralize time querying
Logically, events fired by the event loop run "instantaneously" (and
if they don't, the application is doing work in the main thread, which
it shouldn't). As such, ideally, the only time we want to consult the
actual OS clock is when the event loop wakes up.

Do this, and consolidate all time querying into the event loop.  But
also, for practicality's sake, re-query the current time before
re-entering sleep, to ensure the amount of time to sleep is accurate
despite any slow event handlers.