ae.sys.data: Switch over Data base type from void to ubyte
As per the justification in the comment.

void[] is iffy for raw data, as the language assumes it may have
pointers.  ubyte[], on the other hand, is perfect as the lowest-level
representation of all POD stuff (file formats, protocols, etc.).

We can make this change with minimal breakage thanks to the fact that
void[] has to be cast to something else to be useful anyway, so
existing casts will also apply to the new ubyte[] type of .contents
etc.  As such, we only need to take care of the entry points
(construction, appending, etc.) and add deprecated shims there.