ae.sys.data: Ensure zero-length Data instances don't pin Memory
We have to make a fundamental choice here:

1. Guarantee that a[0..0].contents.ptr != b[0..0].contents.ptr,
   given that a and b are Data with different provenance;
2. Don't pin Memory instances if all Data pointing to it have length 0.

We choose the second option, for practical reasons.

Note that simply keeping the slice while dropping the Memory reference
still doesn't allow us to fulfill the guarantee in (1.), because the
OS may allocate a second instance at the same address as the first.

In general, identity of zero-length types is a long-standing topic in
programming language design.  Note that LDC entirely optimizes away
zero-length static arrays, such that their .ptr is always zero.