ae.net.asockets: Prefer writing to reading
We only handle one event per select call, because the application
logic might change which events we want to accept. So, once we find
one actionable event, we act on it, and continue the event loop.

If a socket is both readable and writable, we therefore used to only
handle the read event. The problem is that if the socket is fully
saturated, the read event will always be called - potentially never
calling the write event.

This can become a problem for echo-server-type applications, where
writes might get queued up indefinitely, with no chance to flush them
to the wire.

For this reason, prefer writing before reading. Writing lets us
un-reference the enqueued data, potentially freeing memory.