ae.utils.time.format: Refactor to avoid circular attribute inference issues
We want to make putToken work with a sink that may or may not be
@safe, but the D compiler isn't smart enough to infer the attribute
of the function because it is recursive.  However, the amount of
recursion is actually fixed - the compiler just can't see that because
we use a runtime "switch" even when the token is known at
compile-time.

Fix this by switching to a different approach of dispatching to the
token-specific formatting logic, which allows selective instantiation
in order to avoid the appearance of unbounded recursion.