ae.utils.main: Allow overriding whether to print the stack trace
The semantics of "debug" aren't quite fully formally defined, so it
could be used to mean anything between "this build has debug
information" to "enable additional debug features that should not
normally be enabled". An additional factor is we can't query whether
debug information is enabled (as there is no corresponding predefined
version for this), which would be useful as it would not be useful to
print stack traces without debug information.

For now, at least allow controlling whether stack traces are printed
or not by adding an optional parameter which allows overriding the
default behavior (which remains driven by the "-debug" flag).