Add AE_DEBUG_FD_LEAK debug feature
This adds a basic debug facility for tracking down file descriptor
leaks. The facility intercepts libc calls which create file
descriptors, such as open, and captures stack traces when each is
created; when a file descriptor above a certain number is allocated,
all stack traces are dumped, clearly showing what is using up all
the file descriptors.

The facility is limited to intercepting calls coming from the main D
application, so it's not as comprehensive as a syscall-based approach,
or even as an LD_PRELOAD-based approach. On the other hand, it is very
easy to use - simply add to your dub.sdl:

dependency "ae:fdtrack" version="..."
debugVersions "AE_DEBUG_FD_LEAK"

and run your application, optionally setting
AE_DEBUG_FD_LEAK_THRESHOLD to the desired threshold (1000 by default).
When the threshold is exceeded, a report will be dumped to the
standard error stream.