ae.utils.promise: Fix except() return type to use CommonType
Change the return type of except() from Promise!(R, F) to
Promise!(CommonType!(Unpromise!R, T), F) to match the behavior of
then(null, onRejected). This ensures that when the rejection handler
returns a type with a common type to T, the result type is correctly
computed.

Add unit test verifying that except() and then(null, ...) produce the
same result type when given equivalent handlers.